refactoring of all controllers

+++ server/controllers/tourController.js
@@ -1,22 +1,20 @@
 const tourService = require("../services/tourService");
 
-
 async function getAllTours(req, reply) {
   try {
     const tours = await tourService.getAllTours();
-    reply.send({ message: "success", tours })
+    reply.send({ message: "success", data: { tours } });
   } catch (error) {
-    reply.status(400).send({ error: error.message });
+    reply.status(400).send({ message: "failure", error: error.message });
   }
 }
-
 
 async function createTour(req, reply) {
   try {
     const tour = await tourService.createTour(req.body);
-    reply.status(201).send({ message: "success", tour });
+    reply.status(201).send({ message: "success", data: { tour } });
   } catch (error) {
-    reply.status(400).send({ error: error.message });
+    reply.status(400).send({ message: "failure", error: error.message });
   }
 }
 
@@ -24,13 +22,13 @@ async function getTour(req, reply) {
   const { id } = req.params;
   try {
     const tour = await tourService.getTourById(id);
-    if (!tour) {
-      reply.status(404).send({ error: "Tour not found" });
-    } else {
-      reply.send({ message: "success", tour });
-    }
+    if (!tour)
+      return reply
+        .status(404)
+        .send({ message: "failure", error: "Tour not found" });
+    reply.send({ message: "success", data: { tour } });
   } catch (error) {
-    reply.status(400).send({ error: error.message });
+    reply.status(400).send({ message: "failure", error: error.message });
   }
 }
 
@@ -38,19 +36,16 @@ async function getTour(req, reply) {
 async function updateTour(req, reply) {
   const { id } = req.params;
   try {
-    const tour = await tourService.getTourById(id);
-    if (!tour) {
-      reply.status(404).send({ error: "Tour not found" });
+    const tourExists = await tourService.getTourById(id);
+    if (!tourExists) {
+      return reply
+        .status(404)
+        .send({ message: "failure", error: "Tour not found" });
     }
-    const updatedTour = await tourService.updateTour(id, req.body);
-    if (!updatedTour) {
-      reply.status(404).send({ error: "Error while updating" });
-    } else {
-      const tour = await tourService.getTourById(id);
-      reply.send({ message: "success", tour })
-    }
+    const tour = await tourService.updateTour(id, req.body);
+    reply.send({ message: "success", data: { tour } });
   } catch (error) {
-    reply.status(400).send({ error: error.message });
+    reply.status(400).send({ message: "failure", error: error.message });
   }
 }
 
@@ -58,14 +53,16 @@ async function updateTour(req, reply) {
 async function deleteTour(req, reply) {
   const { id } = req.params;
   try {
-    const deletedTour = await tourService.deleteTour(id);
-    if (!deletedTour) {
-      reply.status(404).send({ error: "Tour not found" });
-    } else {
-      reply.send({ message: "success", tour })
+    const tourExists = await tourService.getTourById(id);
+    if (!tourExists) {
+      return reply
+        .status(404)
+        .send({ message: "failure", error: "Tour not found" });
     }
+    const tour = await tourService.deleteTour(id);
+    reply.send({ message: "success", data: { tour } });
   } catch (error) {
-    reply.status(400).send({ error: error.message });
+    reply.status(400).send({ message: "failure", error: error.message });
   }
 }
